Meaningful Name Changes
Changing the name of the Funcs.py script to Py2Friends.py as this is actually where the algorithm is housed and it is already generalized. Replacing the old Py2Friends.py with PyFriends_2MRS.py as an example of how the algorithm was run on the 2MRS data set. (one liner).